+++ src/main/java/com/github/jlangch/venice/impl/docgen/DocGenerator.java
@@ -692,6 +692,7 @@ public class DocGenerator {
 		agents.addItem(getDocItem("restart-agent"));
 		agents.addItem(getDocItem("set-error-handler!"));
 		agents.addItem(getDocItem("agent-error"));
+		agents.addItem(getDocItem("await"));
 		agents.addItem(getDocItem("await-for"));
 		agents.addItem(getDocItem("shutdown-agents", false, false));
 		agents.addItem(getDocItem("shutdown-agents?", false, false));