+++ ChangeLog.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### added
 
+- added agents that complement Venice concurrency features. Agents 
+        provide independent, asynchronous change of state.
 - added special form _defonce_. e.g: `(defonce x 100)`
 - added dynamic (thread-local) binding. e.g: `(binding [x 100] (print x))`
 - added _with-out-str_ macro that returns the captured text from stdout.